relocate functions from `clippy_lints::types`
relocate `is_ty_param_lang_item` and `is_ty_param_diagnostic_item` to `clippy_utils`
